Refactored code. Added main page
